Support wildcard/regex for indices param in _remotestore/_restore (#8922)
* Support wildcard/regex for indices param in _remotestore/_restore

(cherry picked from commit 29ebd80fdd6771ccc2023bf432977c38c76bfd55)
